Restructured package structure
