+++ src/main/java/net/evenh/versionmonitor/api/controllers/ProjectController.java
@@ -9,13 +9,13 @@ import net.evenh.versionmonitor.api.exceptions.ProjectNotFoundException;
 import net.evenh.versionmonitor.api.exceptions.SubscriptionAlreadyLinkedToProjectException;
 import net.evenh.versionmonitor.api.exceptions.SubscriptionNotFoundException;
 import net.evenh.versionmonitor.api.exceptions.UnknownProjectTypeException;
-import net.evenh.versionmonitor.application.hosts.HostRegistry;
-import net.evenh.versionmonitor.application.hosts.HostService;
-import net.evenh.versionmonitor.application.projects.AbstractProject;
-import net.evenh.versionmonitor.application.projects.ProjectService;
-import net.evenh.versionmonitor.application.subscriptions.AbstractSubscription;
+import net.evenh.versionmonitor.domain.hosts.HostRegistry;
+import net.evenh.versionmonitor.domain.hosts.HostService;
+import net.evenh.versionmonitor.domain.projects.AbstractProject;
+import net.evenh.versionmonitor.domain.projects.ProjectService;
+import net.evenh.versionmonitor.domain.subscriptions.AbstractSubscription;
 import net.evenh.versionmonitor.application.subscriptions.SubscriptionService;
-import net.evenh.versionmonitor.domain.View;
+import net.evenh.versionmonitor.infrastructure.View;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
