+++ src/main/java/net/evenh/versionmonitor/domain/projects/Project.java
@@ -1,6 +1,6 @@
-package net.evenh.versionmonitor.application.projects;
+package net.evenh.versionmonitor.domain.projects;
 
-import net.evenh.versionmonitor.application.subscriptions.AbstractSubscription;
+import net.evenh.versionmonitor.domain.subscriptions.AbstractSubscription;
 import net.evenh.versionmonitor.domain.releases.Release;
 
 import java.util.List;
