+++ src/main/java/net/evenh/versionmonitor/domain/projects/ProjectRepository.java
@@ -1,4 +1,4 @@
-package net.evenh.versionmonitor.application.projects;
+package net.evenh.versionmonitor.domain.projects;
 
 import net.evenh.versionmonitor.domain.releases.Release;
