+++ src/main/java/net/evenh/versionmonitor/domain/subscriptions/AbstractSubscription.java
@@ -1,9 +1,9 @@
-package net.evenh.versionmonitor.application.subscriptions;
+package net.evenh.versionmonitor.domain.subscriptions;
 
 import com.fasterxml.jackson.annotation.JsonView;
 
 import net.evenh.versionmonitor.api.commands.AddSubscriptionCommand;
-import net.evenh.versionmonitor.domain.View;
+import net.evenh.versionmonitor.infrastructure.View;
 
 import org.hibernate.validator.constraints.NotEmpty;
