+++ src/main/java/net/evenh/versionmonitor/domain/subscriptions/Subscription.java
@@ -1,4 +1,4 @@
-package net.evenh.versionmonitor.application.subscriptions;
+package net.evenh.versionmonitor.domain.subscriptions;
 
 public interface Subscription {
   String getServiceName();
