+++ src/main/java/net/evenh/versionmonitor/infrastructure/config/AsyncConfiguration.java
@@ -1,7 +1,7 @@
 package net.evenh.versionmonitor.infrastructure.config;
 
+import java.util.concurrent.Executor;
 import net.evenh.versionmonitor.infrastructure.config.async.ExceptionHandlingAsyncTaskExecutor;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.aop.interceptor.AsyncUncaughtExceptionHandler;
@@ -14,16 +14,17 @@ import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 
-import java.util.concurrent.Executor;
-
 @Configuration
 @EnableAsync
 @EnableScheduling
 public class AsyncConfiguration implements AsyncConfigurer {
   private final Logger log = LoggerFactory.getLogger(AsyncConfiguration.class);
+  private final VersionmonitorConfiguration properties;
 
   @Autowired
-  private VersionmonitorConfiguration properties;
+  public AsyncConfiguration(VersionmonitorConfiguration properties) {
+    this.properties = properties;
+  }
 
   @Override
   @Bean(name = "taskExecutor")
