+++ src/main/resources/logback-spring.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <configuration scan="true">
-  <include resource="org/springframework/boot/logging/logback/defaults.xml" />
+  <include resource="org/springframework/boot/logging/logback/defaults.xml"/>
 
   <property name="LOG_FILE" value="${LOG_FILE:-${LOG_PATH:-${LOG_TEMP:-${java.io.tmpdir:-/tmp}}}/app.log}"/>
   <property name="LOG_PATTERN" value="%d{ISO8601} %-7level [%35.35thread] %-40.40c: %X %message%n"/>
@@ -31,6 +31,7 @@
 
   <logger name="net.evenh.versionmonitor" level="@logback.loglevel@"/>
   <logger name="org" level="WARN"/>
+  <logger name="sun" level="WARN"/>
 
   <root level="@logback.loglevel@">
     <appender-ref ref="APP_CONSOLE"/>
